Associate Submissions with Users
Ultimately, we want submissions to be tied to the user that submitted
it. For that we are creating a migrations file with
`rails generate migration AddUserToSubmissions user:belongs_to`
and then `rake db:migrate`

Now that we have that association set up in the models, its easy for
us to set up the logic in the controller and views.